LayoutTests/imported/w3c/web-platform-tests/css/css-grid/layout-algorithm/grid-content-distribution-must-account-for-track-sizing-002-expected.txt

XXX XX X XX X XXX

FAIL .grid 1 assert_equals:
<div class="grid justifyContentStart alignContentSpaceBetween" data-expected-width="40" data-expected-height="200">
        <div class="item" data-expected-width="40" data-expected-height="200">XXX XX X XX X XXX</div>
    </div>
width expected 40 but got 80
